Align JavaScript entry files with webpack
